+++ test/integration/messaging/process-publish-message.test.js
@@ -14,15 +14,15 @@ jest.mock('notifications-node-client', () => {
 })
 jest.mock('ffc-messaging')
 const { BlobServiceClient } = require('@azure/storage-blob')
-const config = require('../../app/config/storage')
-const db = require('../../app/data')
-const mockRequest = require('../mocks/request')
-const processPublishMessage = require('../../app/messaging/process-publish-message')
+const config = require('../../../app/config/storage')
+const db = require('../../../app/data')
+const mockRequest = require('../../mocks/request')
+const processPublishMessage = require('../../../app/messaging/process-publish-message')
 const path = require('path')
-const { EMAIL } = require('../../app/methods')
+const { EMAIL } = require('../../../app/methods')
 
 const FILE_NAME = 'FFC_PaymentStatement_SFI_2022_1234567890_2022080515301012.pdf'
-const TEST_FILE = path.resolve(__dirname, '../files/test.pdf')
+const TEST_FILE = path.resolve(__dirname, '../../files/test.pdf')
 
 let blobServiceClient
 let container
